requirements.txt
fastapi[all]
openai
python-dotenv
pydantic==1.*
langchain-community
faiss-cpu
tiktoken
translate
pyairtable